requirements.txt
aws-cdk-lib>=2.232.2,<3.0.0
constructs>=10.0.0,<11.0.0

# Lambda function dependencies (for local testing)
pandas>=2.0.0
pyarrow>=14.0.0
openpyxl>=3.1.0
boto3>=1.34.0
